Split ground truthing into modules and add unit tests.
Break the monolithic ground_truthing __init__.py into focused modules
(app, segments, plotting, session_frames, etc.) and add unit tests
with frame-level assertions where appropriate.
